requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.19.0
certifi>=2023.0.0
requests>=2.31.0
dateparser>=1.1.8
python-dotenv>=1.0.0
